pyproject.toml
[project]
name = "cinereserve"
version = "0.1.0"
description = ""
authors = [
    {name = "Renildo15",email = "renildorabi@ufrn.edu.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "redis (>=7.3.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
